Removed alquimia patch from develop version
The patch has been incorporated into alquimia and is no longer
necessary for the develop version.  The tagged version has not
been updated accordingly and still needs the patch for now.
(When the tagged version gets updated, the patch does need to be
removed from spack altogether, or it will break the build.)