Update cmake.py
Removed line that copies over the user's environment unconditionally.  This should only happen in "dirty" mode.

This line was probably put in at some point trying to debug.  I'm not sure what the effects might be of removing it.  It may turn out that `--dirty` is required for most/all cases (for example, anything using a sysadmin-installed compiler where you need env vars set to have it work).  If this turns out to be the case, we should consider a way for users to set the `--dirty` flag by default for an environment.