config: add initial hardware.yaml schema (#14938)
For cross-compilation, we are adding an optional `hardware.yaml` config file, which can be used to describe the types of target environments available on the system.  Currently this includes the microarchitecture target (`target`) and the OS running on each node (`operating_system`).

For example, on Fugaku, we might use the following hardware.yaml to describe the machine's nodes:

```yaml
hardware:
    nodes:
        - node:
            name: backend
            operating_system: centos8
            target: cascadelake
        - node:
            name: backend2
            operating_system: rhel8
            target: a64fx
```

Other properties, e.g., an optional `sysroot`, or a section describing available networks, can come later.  We expect that this file might also contain hints for setting default values of variants (e.g. you could ask what network fabrics are available on your node and decide how to build `openmpi` based on that. 

The concretizer will use this file to determine what the build environment and the target environment look like, and it will construct build dependencies to run on the host running Spack, and the DAG root + run and link dependencies to run on a selected type of node.  That will come later; this is just the first cut at the schema.